adding tags in examples
