+++ examples/resources/spectrocloud_cluster_gcp/resource.tf
@@ -11,6 +11,7 @@ data "spectrocloud_cluster_profile" "profile" {
 
 resource "spectrocloud_cluster_gcp" "cluster" {
   name             = var.cluster_name
+  tags             = ["dev", "department:devops", "owner:bob"]
   cloud_account_id = data.spectrocloud_cloudaccount_gcp.account.id
 
   cloud_config {
