+++ examples/resources/spectrocloud_team/variables.tf
@@ -2,11 +2,3 @@ variable "sc_host" {}
 variable "sc_username" {}
 variable "sc_password" {}
 variable "sc_project_name" {}
-
-variable "credential_type" {}
-
-variable "aws_access_key" {}
-variable "aws_secret_key" {}
-
-variable "arn" {}
-variable "external_id" {}
